requirements.txt
flask-restful~=0.3.9
flask~=2.2.2
joblib~=1.1.1
jsons~=1.6.3
jsonschema~=4.16.0
numpy~=1.23.4
pandas~=1.5.1
pylint~=2.14.5
pytest~=7.1.2
pytest-flask~=1.2.0
requests~=2.28.1
scikit-learn~=1.1.3
kaleido~=0.2.1
plotly~=5.9.0
matplotlib~=3.6.2
scipy~=1.9.3
python-dotenv~=1.0.1
